backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
yt-dlp
curl_cffi>=0.7.0
python-multipart==0.0.12
python-dotenv==1.0.1
